// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date string, such as 2007-12-03 (YYYY-MM-DD), compliant with ISO 8601 standard for representation of dates using the Gregorian calendar. */
  Date: { input: unknown; output: unknown; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the date-timeformat outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representationof dates and times using the Gregorian calendar. */
  DateTime: { input: unknown; output: unknown; }
  Hex: { input: unknown; output: unknown; }
  /** Raw JSON value */
  Json: { input: unknown; output: unknown; }
  /** The Long scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: { input: unknown; output: unknown; }
  RGBAHue: { input: unknown; output: unknown; }
  RGBATransparency: { input: unknown; output: unknown; }
  /** Slate-compatible RichText AST */
  RichTextAST: { input: unknown; output: unknown; }
};
// ...
export type Collection = Entity & Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<Collection>;
  /** List of Collection versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  posters: Array<Poster>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Entity = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};
// ...
export type LandingHeroImage = Entity & Node & {
  alt: Scalars['String']['output'];
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Get the document in other stages */
  documentInStages: Array<LandingHeroImage>;
  /** List of LandingHeroImage versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  image: Asset;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars['ID']['output'];
  /** The Stage of an object */
  stage: Stage;
};

export type Orientation =
  | 'horizontal'
  | 'vertical';
// ...
export type Poster = Entity & Node & {
  collection?: Maybe<Collection>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  currentPrice: Scalars['Int']['output'];
  description?: Maybe<Scalars['String']['output']>;
  /** Get the document in other stages */
  documentInStages: Array<Poster>;
  /** List of Poster versions */
  history: Array<Version>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  images: Array<Asset>;
  orientation: Orientation;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  scheduledIn: Array<ScheduledOperation>;
  slug: Scalars['String']['output'];
  /** System stage field */
  stage: Stage;
  title: Scalars['String']['output'];
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type ScheduledOperation = Entity & Node & {
  affectedDocuments: Array<ScheduledOperationAffectedDocument>;
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** User that created this document */
  createdBy?: Maybe<User>;
  /** Operation description */
  description?: Maybe<Scalars['String']['output']>;
  /** Get the document in other stages */
  documentInStages: Array<ScheduledOperation>;
  /** Operation error message */
  errorMessage?: Maybe<Scalars['String']['output']>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** User that last published this document */
  publishedBy?: Maybe<User>;
  /** Raw operation payload including all details, this field is subject to change */
  rawPayload: Scalars['Json']['output'];
  /** The release this operation is scheduled for */
  release?: Maybe<ScheduledRelease>;
  /** System stage field */
  stage: Stage;
  /** operation Status */
  status: ScheduledOperationStatus;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
  /** User that last updated this document */
  updatedBy?: Maybe<User>;
};
// ...
export type Stage =
  /** The Draft is the default stage for all your content. */
  | 'DRAFT'
  /** The Published stage is where you can publish your content to. */
  | 'PUBLISHED';

export type SystemDateTimeFieldVariation =
  | 'BASE'
  | 'COMBINED'
  | 'LOCALIZATION';

export type UnpublishLocaleInput = {
  /** Locales to unpublish */
  locale: Locale;
  /** Stages to unpublish selected locales from */
  stages: Array<Stage>;
};

/** User system model */
export type User = Entity & Node & {
  /** The time the document was created */
  createdAt: Scalars['DateTime']['output'];
  /** Get the document in other stages */
  documentInStages: Array<User>;
  /** The unique identifier */
  id: Scalars['ID']['output'];
  /** Flag to determine if user is active or not */
  isActive: Scalars['Boolean']['output'];
  /** User Kind. Can be either MEMBER, PAT or PUBLIC */
  kind: UserKind;
  /** The username */
  name: Scalars['String']['output'];
  /** Profile Picture url */
  picture?: Maybe<Scalars['String']['output']>;
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  /** System stage field */
  stage: Stage;
  /** The time the document was updated */
  updatedAt: Scalars['DateTime']['output'];
};
// ...
export type Version = {
  createdAt: Scalars['DateTime']['output'];
  id: Scalars['ID']['output'];
  revision: Scalars['Int']['output'];
  stage: Stage;
};
// ...
export type CollectionItemFragment = { name: string };

export type CollectionsGetQueryVariables = Exact<{
  take: Scalars['Int']['input'];
}>;


export type CollectionsGetQuery = { collections: Array<{ name: string }> };

export type LandingHeroImageItemFragment = { id: string, alt: string, image: { url: string } };

export type LandingHeroImagesGetQueryVariables = Exact<{
  width?: InputMaybe<Scalars['Int']['input']>;
  height?: InputMaybe<Scalars['Int']['input']>;
}>;


export type LandingHeroImagesGetQuery = { landingHeroImages: Array<{ id: string, alt: string, image: { url: string } }> };

export type PosterGetBySlugQueryVariables = Exact<{
  slug: Scalars['String']['input'];
}>;


export type PosterGetBySlugQuery = { poster?: { id: string, title: string, currentPrice: number, slug: string, description?: string | null, orientation: Orientation, images: Array<{ url: string }> } | null };

export type PosterItemFragment = { id: string, title: string, currentPrice: number, slug: string, description?: string | null, orientation: Orientation, images: Array<{ url: string }> };

export type PostersGetByOrientationQueryVariables = Exact<{
  take: Scalars['Int']['input'];
  orientation?: InputMaybe<Orientation>;
}>;


export type PostersGetByOrientationQuery = { posters: Array<{ id: string, title: string, currentPrice: number, slug: string, description?: string | null, orientation: Orientation, images: Array<{ url: string }> }> };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
export const CollectionItemFragmentDoc = new TypedDocumentString(`
    fragment CollectionItem on Collection {
  name
}
    `, {"fragmentName":"CollectionItem"}) as unknown as TypedDocumentString<CollectionItemFragment, unknown>;
export const LandingHeroImageItemFragmentDoc = new TypedDocumentString(`
    fragment LandingHeroImageItem on LandingHeroImage {
  id
  image {
    url(
      transformation: {image: {resize: {width: $width, height: $height, fit: clip}}}
    )
  }
  alt
}
    `, {"fragmentName":"LandingHeroImageItem"}) as unknown as TypedDocumentString<LandingHeroImageItemFragment, unknown>;
export const PosterItemFragmentDoc = new TypedDocumentString(`
    fragment PosterItem on Poster {
  id
  title
  images {
    url
  }
  currentPrice
  slug
  description
  orientation
}
    `, {"fragmentName":"PosterItem"}) as unknown as TypedDocumentString<PosterItemFragment, unknown>;
export const CollectionsGetDocument = new TypedDocumentString(`
    query CollectionsGet($take: Int!) {
  collections(first: $take) {
    ...CollectionItem
  }
}
    fragment CollectionItem on Collection {
  name
}`) as unknown as TypedDocumentString<CollectionsGetQuery, CollectionsGetQueryVariables>;
export const LandingHeroImagesGetDocument = new TypedDocumentString(`
    query LandingHeroImagesGet($width: Int, $height: Int) {
  landingHeroImages {
    ...LandingHeroImageItem
  }
}
    fragment LandingHeroImageItem on LandingHeroImage {
  id
  image {
    url(
      transformation: {image: {resize: {width: $width, height: $height, fit: clip}}}
    )
  }
  alt
}`) as unknown as TypedDocumentString<LandingHeroImagesGetQuery, LandingHeroImagesGetQueryVariables>;
export const PosterGetBySlugDocument = new TypedDocumentString(`
    query PosterGetBySlug($slug: String!) {
  poster(where: {slug: $slug}) {
    ...PosterItem
  }
}
    fragment PosterItem on Poster {
  id
  title
  images {
    url
  }
  currentPrice
  slug
  description
  orientation
}`) as unknown as TypedDocumentString<PosterGetBySlugQuery, PosterGetBySlugQueryVariables>;
export const PostersGetByOrientationDocument = new TypedDocumentString(`
    query PostersGetByOrientation($take: Int!, $orientation: Orientation) {
  posters(first: $take, where: {orientation: $orientation}) {
    ...PosterItem
  }
}
    fragment PosterItem on Poster {
  id
  title
  images {
    url
  }
  currentPrice
  slug
  description
  orientation
}`) as unknown as TypedDocumentString<PostersGetByOrientationQuery, PostersGetByOrientationQueryVariables>;